bp/bpdata/requirements.txt
python3
WebSocketclient
sqlalchemy
python-binance
btfxwss
fcoin
apscheduler
